# apps/server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stationary Server API",
    "description": "All API endpoints for the Stationary backend service, supporting export to tools like Apifox / Postman.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9400",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/user/": {
      "get": {
        "summary": "Get current logged-in user info",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string",
                          "nullable": true
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/user/tokens": {
      "post": {
        "summary": "Create API access token (Token)",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "prefix": {
                          "type": "string"
                        },
                        "first_four": {
                          "type": "string"
                        },
                        "last_four": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "library_id": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "expires_at": {
                          "type": "string",
                          "nullable": true
                        },
                        "create_time": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "library_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "expires_in_seconds": {
                    "type": "integer",
                    "minimum": 1,
                    "nullable": true
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all API tokens of the current user",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "prefix": {
                            "type": "string"
                          },
                          "first_four": {
                            "type": "string"
                          },
                          "last_four": {
                            "type": "string"
                          },
                          "library_id": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "last_used_at": {
                            "type": "string",
                            "nullable": true
                          },
                          "expires_at": {
                            "type": "string",
                            "nullable": true
                          },
                          "create_time": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/user/tokens/{id}": {
      "delete": {
        "summary": "Revoke specified API token",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authenticate requests using API token (Authorization: Bearer <token>)"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Authenticate requests using Better Auth session cookie"
      }
    }
  }
}